Added modal for knowledge graph
