+++ src/App.tsx
@@ -1,11 +1,5 @@
 import React from "react";
-import {
-  Error404Page,
-  AboutPage,
-  PdfViewerPage,
-  HomePage,
-  TestGraphPage,
-} from "./pages";
+import { Error404Page, AboutPage, PdfViewerPage, HomePage } from "./pages";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { Header } from "./components";
 import "./styles/main.scss";
@@ -18,7 +12,6 @@ function App() {
         <Route path="/" element={<HomePage />} />
         <Route path="/pdfviewer/" element={<PdfViewerPage />} />
         <Route path="/about" element={<AboutPage />} />
-        <Route path="/testGraph" element={<TestGraphPage />} />
         <Route path="*" element={<Error404Page />} />
       </Routes>
     </Router>
