+++ src/components/KnowledgeGraph/KnowledgeGraph.types.ts
@@ -1,3 +1,5 @@
+import { Dispatch, SetStateAction } from "react";
+
 export type KnowledgeGraphProps = {
   /**
    * Id prefix for DOM elements inside KnowledgeGraph component
@@ -12,4 +14,8 @@ export type KnowledgeGraphProps = {
    */
   className?: string;
   paperTitle?: string;
+  setIsVisible: Dispatch<SetStateAction<boolean>>;
+  setElements: Dispatch<
+    SetStateAction<Array<Record<string, Record<string, string>>> | null>
+  >;
 };
