+++ src/components/KnowledgeGraphModal/cy-style.json
