+++ src/components/KnowledgeGraphModal/index.ts
@@ -1,0 +1,2 @@
+export { default as KnowledgeGraphModal } from "./KnowledgeGraphModal";
+export type { KnowledgeGraphModalProps } from "./KnowledgeGraphModal.types";
