--- src/pages/TestGraphPage/TestGraphPage.scss
@@ -1,2 +1,0 @@
-.sa-test-graph-page {
-}
