--- src/pages/TestGraphPage/index.ts
@@ -1,2 +1,0 @@
-export { default as TestGraphPage } from "./TestGraphPage";
-export type { TestGraphPageProps } from "./TestGraphPage.types";
