+++ src/pages/index.ts
@@ -1,5 +1,4 @@
 /* PLOP_INJECT_IMPORT */
-export * from "./TestGraphPage";
 export * from "./HomePage";
 export * from "./AboutPage";
 export * from "./PdfViewerPage";
