+++ src/shared/getDataForKnowledgeGraph.tsx
@@ -11,9 +11,10 @@ export const formatDataForDisplay = (data: Record<string, string>) => {
     returnValues["Number of References"] = data.referenceCount;
   if (data.citationCount != null)
     returnValues["Number of Citations"] = data.citationCount;
-  if (data.published != null)
+  if (data.published != null && data.published.length > 0)
     returnValues["Date of Publication"] = data.published;
-  if (data.journal != null) returnValues["Published Journal"] = data.journal;
+  if (data.journal != null && data.journal.length > 0)
+    returnValues["Published Journal"] = data.journal;
 
   return returnValues;
 };
