+++ src/styles/main.scss
@@ -3,6 +3,7 @@
 @import "./theme/typography.scss";
 
 /* PLOP_INJECT_IMPORT */
+@import '../components/KnowledgeGraphModal/KnowledgeGraphModal.scss';
 @import "../components/KnowledgeGraph/KnowledgeGraph.scss";
 @import "../components/TitleClose/TitleClose.scss";
 @import "../components/CustomSummary/CustomSummary.scss";
