+++ src/styles/theme/colors.scss
@@ -26,6 +26,9 @@
 }
 
 /* PLOP_INJECT_IMPORT */
+.sa-knowledge-graph-modal{
+}
+
 .sa-knowledge-graph {
 }
